+++ lib/macaw_framework.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "macaw_framework/endpoint_not_mapped_error"
+require_relative "macaw_framework/request_data_filtering"
 require_relative "macaw_framework/http_status_code"
 require_relative "macaw_framework/version"
 require "socket"
@@ -12,14 +13,17 @@ module MacawFramework
   # starting the web server.
   class Macaw
     include(HttpStatusCode)
-    def initialize
+    ##
+    # @param {Logger} custom_log
+    def initialize(custom_log = nil)
       begin
-        config = JSON.parse(File.read("application.json"))
-        @port = config["macaw"]["port"]
+        config = JSON.parse(File.read('application.json'))
+        @port = config['macaw']['port']
       rescue StandardError
         @port ||= 8080
       end
       @port ||= 8080
+      @macaw_log ||= custom_log.nil? ? Logger.new($stdout) : custom_log
     end
 
     ##
@@ -29,8 +33,9 @@ module MacawFramework
     # @param {Proc} block
     # @return {Integer, String}
     def get(path, &block)
-      path_clean = path[0] == "/" ? path[1..].gsub("/", "_") : path.gsub("/", "_")
-      define_singleton_method("get_#{path_clean}", block)
+      path_clean = RequestDataFiltering.extract_path(path)
+      @macaw_log.info("Defining GET endpoint at #{path_clean}")
+      map_new_endpoint('get', path_clean, &block)
     end
 
     ##
@@ -40,8 +45,9 @@ module MacawFramework
     # @param {Proc} block
     # @return {String, Integer}
     def post(path, &block)
-      path_clean = path[0] == "/" ? path[1..].gsub("/", "_") : path.gsub("/", "_")
-      define_singleton_method("post_#{path_clean}", block)
+      path_clean = path[0] == '/' ? path[1..].gsub('/', '_') : path.gsub('/', '_')
+      @macaw_log.info("Defining POST endpoint at #{path_clean}")
+      map_new_endpoint('post', path_clean, &block)
     end
 
     ##
@@ -51,8 +57,9 @@ module MacawFramework
     # @param {Proc} block
     # @return {String, Integer}
     def put(path, &block)
-      path_clean = path[0] == "/" ? path[1..].gsub("/", "_") : path.gsub("/", "_")
-      define_singleton_method("put_#{path_clean}", block)
+      path_clean = path[0] == '/' ? path[1..].gsub('/', '_') : path.gsub('/', '_')
+      @macaw_log.info("Defining PUT endpoint at #{path_clean}")
+      map_new_endpoint('put', path_clean, &block)
     end
 
     ##
@@ -62,8 +69,9 @@ module MacawFramework
     # @param {Proc} block
     # @return {String, Integer}
     def patch(path, &block)
-      path_clean = path[0] == "/" ? path[1..].gsub("/", "_") : path.gsub("/", "_")
-      define_singleton_method("patch_#{path_clean}", block)
+      path_clean = path[0] == '/' ? path[1..].gsub('/', '_') : path.gsub('/', '_')
+      @macaw_log.info("Defining PATCH endpoint at #{path_clean}")
+      map_new_endpoint('patch', path_clean, &block)
     end
 
     ##
@@ -73,67 +81,48 @@ module MacawFramework
     # @param {Proc} block
     # @return {String, Integer}
     def delete(path, &block)
-      path_clean = path[0] == "/" ? path[1..].gsub("/", "_") : path.gsub("/", "_")
-      define_singleton_method("delete_#{path_clean}", block)
+      path_clean = path[0] == '/' ? path[1..].gsub('/', '_') : path.gsub('/', '_')
+      @macaw_log.info("Defining DELETE endpoint at #{path_clean}")
+      map_new_endpoint('delete', path_clean, &block)
     end
 
     ##
     # Starts the web server
     def start!
+      @macaw_log.info("Starting server at port #{@port}")
+      time = Time.now
       server = TCPServer.open(@port)
-      puts "Starting server at port #{@port}"
+      @macaw_log.info("Server started in #{Time.now - time} seconds.")
       loop do
         Thread.start(server.accept) do |client|
-          client.select
-          method_name, headers, body = extract_client_info(client)
+          path, method_name, headers, body, parameters = RequestDataFiltering.extract_client_info(client)
           raise EndpointNotMappedError unless respond_to?(method_name)
 
-          message, status = send(method_name, headers, body)
+          @macaw_log.info("Running #{path.gsub("\n", '').gsub("\r", '')}")
+          message, status = send(method_name, headers, body, parameters)
           status ||= 200
-          message ||= "Ok"
+          message ||= 'Ok'
           client.puts "HTTP/1.1 #{status} #{HTTP_STATUS_CODE_MAP[status]} \r\n\r\n#{message}"
           client.close
         rescue EndpointNotMappedError
           client.print "HTTP/1.1 404 Not Found\r\n\r\n"
           client.close
-        rescue StandardError
+        rescue StandardError => e
           client.print "HTTP/1.1 500 Internal Server Error\r\n\r\n"
+          @macaw_log.info("Error: #{e}")
           client.close
         end
       end
     rescue Interrupt
-      puts "Macaw stop flying for some seeds..."
+      @macaw_log.info('Stopping server')
+      server.close
+      @macaw_log.info('Macaw stop flying for some seeds...')
     end
 
     private
 
-    ##
-    # Method responsible for extracting information
-    # provided by the client like Headers and Body
-    def extract_client_info(client)
-      method_name = client.gets.gsub("HTTP/1.1", "").gsub("/", "_").strip!.downcase
-      method_name.gsub!(" ", "")
-      body_first_line, headers = extract_headers(client)
-      body = extract_body(client, body_first_line, headers["Content-Length"].to_i)
-      [method_name, headers, body]
-    end
-
-    ##
-    # Extract application headers
-    def extract_headers(client)
-      header = client.gets.delete("\n").delete("\r")
-      headers = {}
-      while header.match(%r{[a-zA-Z0-9\-/*]*: [a-zA-Z0-9\-/*]})
-        split_header = header.split(":")
-        headers[split_header[0]] = split_header[1][1..]
-        header = client.gets.delete("\n").delete("\r")
-      end
-      [header, headers]
-    end
-
-    def extract_body(client, body_first_line, content_length)
-      body = client.read(content_length)
-      body_first_line << body.to_s
+    def map_new_endpoint(prefix, path, &block)
+      define_singleton_method("#{prefix}_#{path}", block)
     end
   end
 end