+++ CHANGELOG.md
@@ -7,3 +7,10 @@
 ## [0.1.1] - 2022-12-06
 
 - Adding support for headers and body
+
+## [0.1.2] - 2022-12-10
+
+- Adding support to URL parameters
+- Adding logs to the framework activity
+- Removing undefined Status Codes from http_status_code hash
+- Moving methods from Macaw class to RequestDataFiltering module, respecting SOLID